+++ persistence/querier.go
@@ -9,9 +9,9 @@ type Querier interface {
 	Close()
 
 	SaveDevice(device domain.Device) error
-	GetDeviceById(id uuid.UUID) (*domain.Device, error)
+	GetDevices() ([]domain.Device, error)
+	GetDevice(id uuid.UUID) (*domain.Device, error)
 	UpdateDevice(device domain.Device) error
-
 	SaveSignature(device domain.Signature) (uuid.UUID, error)
-	GetSignaturesByDeviceId(id uuid.UUID) ([]domain.Signature, error)
+	GetSignaturesByDevice(id uuid.UUID) ([]domain.Signature, error)
 }